requirements.txt
SQLAlchemy==1.4.46
psycopg2-binary==2.9.5
python-dotenv==0.21.1
python-whois==0.8.0
alembic==1.9.2
asyncio==3.4.3
aiohttp==3.10.1
pytest==7.3.1
pytest-asyncio==0.21.0
sqlalchemy_utils==0.41.1
